backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
bcrypt==4.1.2
pyjwt==2.8.0
pytest==7.4.3
hypothesis==6.92.1
python-dotenv==1.0.0
